requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coloredlogs==14.0
discord.py==1.3.2
humanfriendly==8.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
PyNaCl==1.3.0
pyreadline==2.1
pytube3==9.6.4
six==1.14.0
typing-extensions==3.7.4.1
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
youtube-dl==2020.3.8
